refactor: 프로젝트 토글 각각 분리
