+++ src/assets/icons/index.ts
@@ -4,5 +4,8 @@ import NextArrow from './arrow_right.svg?react';
 import GreenArrow from './arrow_right_green.svg?react';
 import RedArrow from './arrow_right_red.svg?react';
 import UpArrow from './arrow_up.svg?react';
+import ScenarioLogo from './file_branch.svg?react';
+import DashboardLogo from './grid.svg?react';
+import InformationLogo from './info.svg?react';
 
-export { DownArrow, GreenArrow, NextArrow, PreArrow, RedArrow, UpArrow };
+export { DashboardLogo, DownArrow, GreenArrow, InformationLogo, NextArrow, PreArrow, RedArrow, ScenarioLogo, UpArrow };
