+++ src/components/common/sidebar/projectItems/projectItems.style.ts
