+++ src/components/common/sidebar/projectList/projectList.tsx
@@ -2,9 +2,8 @@ import { useDispatch } from '@/hooks/common/useCustomRedux.ts';
 import useProjectList from '@/hooks/sidebar/sidebar';
 
 import { MODAL_TYPES } from '@/components/common/modalProvider/modalProvider.tsx';
+import ProjectItems from '@/components/common/sidebar/projectItems/projectItems';
 import * as S from '@/components/common/sidebar/projectList/projectList.style';
-
-import ProjectItem from '../projectItem/ProjectItem';
 
 import Plus from '@/assets/icons/add.svg?react';
 import { openModal } from '@/slices/modalSlice.ts';
@@ -20,12 +19,13 @@ export default function ProjectList() {
       <S.Projects className="menu">
         <S.ProjectText>Projects</S.ProjectText>
         <Plus
+          role="button"
           onClick={() => {
             modalDispatch(openModal({ modalType: MODAL_TYPES.CreateProjectModal, modalProps: { projectLength: projectList?.result.projectList.length } }));
           }}
         />
       </S.Projects>
-      <ProjectItem projects={projects} />
+      <ProjectItems projects={projects} />
     </>
   );
 }
