+++ src/constants/projectToggle/projectToggle.ts
@@ -1,0 +1,16 @@
+import { DashboardLogo, InformationLogo, ScenarioLogo } from '@/assets/icons';
+
+export const projectItem = [
+  {
+    name: 'Dashboard',
+    svg: DashboardLogo,
+  },
+  {
+    name: 'Information',
+    svg: InformationLogo,
+  },
+  {
+    name: 'Scenario',
+    svg: ScenarioLogo,
+  },
+];
